1. Moved the field-name textfield to the home page 2. Added a description on the home page about the user history feature 3. Removed the accordion and combined it to the map box
