+++ src/components/Home/index.jsx
@@ -5,9 +5,12 @@ import {
   Box, Card, Stack, ToggleButton, ToggleButtonGroup, Typography, styled,
 } from '@mui/material';
 import { useNavigate } from 'react-router-dom';
+import { PSATextField } from 'shared-react-components/src';
 import { get, set } from '../../store/Store';
+import Help from '../../shared/Help';
 import NavButton from '../../shared/Navigate/NavButton';
 import About from '../About';
+import { historyStates } from '../../store/inits';
 
 const BiomassMethodButton = styled(ToggleButton)(() => ({
   '&.Mui-selected': {
@@ -38,6 +41,17 @@ const Home = () => {
   const dispatch = useDispatch();
   // const privacy = useSelector(get.privacy);
   const biomassCalcMode = useSelector(get.biomassCalcMode);
+
+  const { historyState, userHistoryList } = useSelector(get.user);
+  const field = useSelector(get.field);
+  const [fieldName, setFieldName] = useState(field);
+
+  // Check if field name exists in user history
+  const isFieldNameExisted = () => {
+    if (historyState === historyStates.imported) return false;
+    const result = userHistoryList.find((history) => history.label === 'history-'.concat(fieldName));
+    return result !== undefined;
+  };
 
   useEffect(() => {
     if (window.location.toString().includes('PSA')) {
@@ -76,6 +90,42 @@ const Home = () => {
           </Typography>
         </Box>
       </Stack>
+      <Box
+        sx={{
+          marginTop: 3,
+          display: 'flex',
+          justifyContent: 'center',
+          alignItems: 'center',
+        }}
+      >
+        <Box
+          sx={{
+            backgroundColor: '#fff',
+            padding: 4,
+            borderRadius: 2,
+            boxShadow: 5,
+            textAlign: 'center',
+          }}
+        >
+          <Typography variant="h6" mb={2}>
+            Would you like to save your selection history? Simply give it a name, and your selections will be stored after you've made all your selections. You can easily retrieve them later from the dropdown in the header.
+          </Typography>
+          <PSATextField
+            label="Name your Field (optional)"
+            value={fieldName}
+            onChange={(e) => setFieldName(e.target.value)}
+            error={isFieldNameExisted()}
+            helperText={isFieldNameExisted() ? 'Field name existed!' : null}
+            onBlur={() => {
+              if (!isFieldNameExisted()) dispatch(set.field(fieldName));
+            }}
+            onKeyDown={(e) => {
+              if (e.key === 'Enter' && !isFieldNameExisted()) dispatch(set.field(fieldName));
+            }}
+          />
+          <Help />
+        </Box>
+      </Box>
       <Box sx={{ height: 60 }} />
       <Stack spacing={2} direction="column">
         <Stack justifyContent="space-around" alignItems="center" sx={{ flexDirection: { sm: 'column', md: 'row' } }}>
