+++ src/components/Location/index.jsx
@@ -2,15 +2,13 @@ import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import { useDispatch, useSelector } from 'react-redux';
 import { Box, Stack, Typography, Badge, Paper, Grid } from '@mui/material';
-import { PSALoadingSpinner, PSATooltip, PSAAccordion, PSATextField } from 'shared-react-components/src';
+import { PSALoadingSpinner, PSATooltip } from 'shared-react-components/src';
 import BiomassMap from '../../shared/Map/BiomassMap';
-import Help from '../../shared/Help';
 import { get, set } from '../../store/Store';
 import NavButton from '../../shared/Navigate/NavButton';
 import useFetchHLS from '../../hooks/useFetchHLS';
 import Datebox from '../../shared/BiomassData/Datebox';
 import { AreaErrorModal, TaskFailModal } from '../../shared/BiomassData/Warnings';
-import { historyStates } from '../../store/inits';
 
 const nextButtonBadgeContent = () => (
   <PSATooltip title="No polygon is drawn" tooltipContent={(
@@ -28,77 +26,38 @@ const Location = ({ barebone = false }) => {
   const biomassTaskResults = useSelector(get.biomassTaskResults);
   const polyDrawTooBig = useSelector(get.polyDrawTooBig);
   const biomassFetchIsFailed = useSelector(get.biomassFetchIsFailed);
-  const field = useSelector(get.field);
-  const { historyState, userHistoryList } = useSelector(get.user);
-
-  const [fieldName, setFieldName] = useState(field);
-  const [isExpanded, setIsExpanded] = useState(false); // State for controlling accordion expansion
 
   // API for getting biomass map
   useFetchHLS();
 
-  // Check if field name exists in user history
-  const isFieldNameExisted = () => {
-    if (historyState === historyStates.imported) return false;
-    const result = userHistoryList.find((history) => history.label === 'history-'.concat(fieldName));
-    return result !== undefined;
-  };
-
   return (
     <Box sx={{ width: '100%', padding: '0rem' }}>
-      <Box mb={-2}>
-        <PSAAccordion
-          expanded={isExpanded} // Controlled by state
-          onChange={() => setIsExpanded(!isExpanded)} // Toggle expanded state
-          summaryContent={!barebone && (
+      <Box sx={{ margin: '2rem 0rem' }}>
+        <Paper sx={{ padding: '1rem', borderRadius: '1rem' }}>
+          {!barebone && (
             <Typography variant="h5" gutterBottom>
               Where is your Field located?
             </Typography>
           )}
-          detailsContent={(
-            <>
-              <Stack mb={1}>
-                <Typography variant="h8" gutterBottom>
-                  Enter your address or zip code to determine your field&apos;s location. You can then zoom in and click to pinpoint it on the map. If
-                  you know your exact coordinates, you can enter them in search bar separated by comma (ex. 37.7, -80.2 ).
-                </Typography>
-                {isSatelliteMode && (
-                  <Typography variant="h8" gutterBottom pt={1}>
-                    Specify your field&apos;s boundary on the map using the drawing tool.
-                  </Typography>
-                )}
-              </Stack>
-              <Box mb={2}>
-              <PSATextField
-                label="Name your Field (optional)"
-                value={fieldName}
-                onChange={(e) => setFieldName(e.target.value)}
-                error={isFieldNameExisted()}
-                helperText={isFieldNameExisted() ? 'Field name existed!' : null}
-                onBlur={() => {
-                  if (!isFieldNameExisted()) dispatch(set.field(fieldName));
-                }}
-                onKeyDown={(e) => {
-                  if (e.key === 'Enter' && !isFieldNameExisted()) dispatch(set.field(fieldName));
-                }}
-              />
-                <Help />
-              </Box>
-              <Stack mt={5} gap={1}>
-                {isSatelliteMode && (
-                  <Typography variant="h8" gutterBottom>
-                    Specify your crop&apos;s planting and termination dates.
-                  </Typography>
-                )}
-                <Datebox />
-              </Stack>
-            </>
-          )}
-          testId="location-accordion"
-        />
-      </Box>
-      <Box sx={{ margin: '2rem 0rem' }}>
-        <Paper sx={{ padding: '1rem', borderRadius: '1rem' }}>
+          <Stack mb={1}>
+            <Typography variant="h8" gutterBottom>
+              Enter your address or zip code to determine your field&apos;s location. You can then zoom in and click to pinpoint it on the map. If
+              you know your exact coordinates, you can enter them in search bar separated by comma (ex. 37.7, -80.2 ).
+            </Typography>
+            {isSatelliteMode && (
+              <Typography variant="h8" gutterBottom pt={1}>
+                Specify your field&apos;s boundary on the map using the drawing tool.
+              </Typography>
+            )}
+          </Stack>
+          <Stack mt={5} gap={1} margin={5}>
+            {isSatelliteMode && (
+              <Typography variant="h8" gutterBottom>
+                Specify your crop&apos;s planting and termination dates.
+              </Typography>
+            )}
+            <Datebox />
+          </Stack>
           {biomassFetchIsLoading && (
             <Box>
               <Grid
